+++ src/main/kotlin/icu/windea/pls/ai/requests/PlsAiPolishLocalisationsRequest.kt
@@ -1,13 +1,14 @@
 package icu.windea.pls.ai.requests
 
+import com.intellij.openapi.project.Project
 import com.intellij.psi.*
 import icu.windea.pls.lang.*
 import icu.windea.pls.localisation.psi.*
-import org.apache.tools.ant.*
 
-class PlsAiColorizeLocalisationsRequest(
+class PlsAiPolishLocalisationsRequest(
     val localisations: List<ParadoxLocalisationProperty>,
-    val description: String?,
+    val text: String,
+    val description: String?, //TODO 2.0.0-dev 可以自定义的润色要求
     val file: PsiFile?,
     val project: Project
 ) : PlsAiRequest {