+++ src/main/kotlin/icu/windea/pls/ai/services/PlsAiService.kt
@@ -1,6 +1,3 @@
 package icu.windea.pls.ai.services
 
-import dev.langchain4j.service.*
-import icu.windea.pls.model.*
-
 interface PlsAiService